+++ CHANGELOG.md
@@ -9,3 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.1.0] - 2018-06-28
 ### Added
 - Endpoints working without best algorithm performance
+
+## [0.1.1] - 2018-06-29
+### Fix
+- Minimum always returning 0 for transactions with positive amount
+
+## [0.1.2] - 2018-06-29
+### Fix
+- Maximum always returning 0 for transactions with negative amount